services/knowledge-service/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-service"
version = "0.1.0"
description = "Knowledge service - document ingestion, chunking, embedding, and vector search"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "openai>=1.30.0",
    "boto3>=1.34.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.7.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "va-common",
]

[tool.setuptools.packages.find]
where = ["src"]
